backend/requirements.txt
fastapi
uvicorn
python-multipart

pytesseract
opencv-python
Pillow
numpy
PyMuPDF

sentence-transformers
faiss-cpu


requests
python-dotenv
